Legacy commit
This is the complete road up to the current state of this program. (animate_diffims_v4.py)
The different versions of animate_diffims show how the program has changed over time, with a short description of the changes in each version at the top of the file.
get_late_abmxfrc.py is one of the early tests for automating the download of ZTF images.
Lastly, query_list.txt is a list of options set for each run of this program.

Note that all of this was made on 1 specific computer, and the location to find / store data is set for that computer. If one wishes to use one of these on a different computers, the references to those specific locations should be changed.