Restructured version
animate_diffims_v5.py is a completely rewritten version of the code that addresses several issues at once.
It has been written with the PEP8 style guide in mind. (or at the very least an attempt was made)

Changes / improvements were made in the following areas:
- Better control section. Removed cluttering comments, and several options now have default value.
- An anim_target class. Easy to keep all relevant information about the target in one place.
- More comments + structured comments. A well documented code increases readability.
- Controllable subplots. Each subplot now has its own init and update function, which allows to cherry pick the wanted subplots in each animation. This also allows the addition of custom subplots if desired.
- Bug fixes:
    - Fixed the misalignment of the images.
    - Fixed an issue where the images were not rotated when needed.
    - Added some lines for increased readability.
- And more (?). Basically putting this here in case I forgot to mention something.

Anyway, this new version should make the program easier to use, and a lot clearer to adapt to everyone's specific needs.
As always, have fun.