1st version of a description of the program
This is a first version / attempt at trying to write an explanation on how to use this program.
Included are a basic description of the settings and where to find them, as well as how to change them. A brief description on how to add a custom subplot to the animation is given as well.

Note that I wrote both the code and the description, and as such there exists a possibility that the explanation here is lacking because something feels 'obvious' for me and I forgot to include it, or my written way of thinking is difficult for others to follow.  This will be corrected once those issues are brought to my attention.