error fixes
Found some errors when using the code, I had to change it to this to make things work. A part of the problem was that the object was very close to the edge of the image, and increasing the min_visible region fixed that. This however prompted an issue where a reference image would be discarded for this reason, which I don't necessarily want. Also changed the max value in the cutout from np.max() to np.nanmax() due to weird behaviour when NaNs are involved.