Spectra plots first version
This is a first version of functions for showing spectra during the animation. Though it works as intended I am not yet completely happy about the way I implemented it, hence the new branch.

The current input is a csv file with the following columns: fileloc, instrument, mjd, wl_min, wl_max. The spectra are sorted on mjd and shown above the light curve, in the same place as the wireframe and text plots. I had to do it this way since the spectra I used in the test case came from different sources, and retrieving some of the data has proven difficult if it is stored differently for the different files (especially intstrument & mjd). I am not sure if I want to keep it this way, or change the input.

The horizontal axis is set by the minimum value in wl_min and maximum value in wl_max to ensure the full wavelength range is shown for all spectra without changing the scale. If you want a buffer of e.g. minimal 5nm on each side it has to be changed in the csv file. The vertical axis is set as an argument with default value, again to make sure the scales aren't changed. The spectra themselves are shown normalised.

Starting when the difference images are shown, each frame a function checks if the next spectrum is taken before the next image. If no, change nothing. If yes, show the next spectrum, and show the mjd and instrument in the title. This way, if there are multiple spectra taken between two images, no spectrum is skipped. (Assuming there are enough images to show as well. If there are two spectra taken between the second to last and last images, the first of those two spectra will be shown with the last image, and the last spectrum won't be shown as there are no more images left to show it with.)

Additionally, grey dotted vertical lines are shown in the locations of the spectra in the light curve plot to easily see when they are taken in comparison with the lc points. Might want to consider highlighting the once currently being shown (by making it a dark yellow / gold or brown colour or something like that).